pyproject.toml
[project]
name = "bduq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
